+++ console/migrations/m220916_205357_add_price_fields_to_game_table.php
@@ -12,8 +12,8 @@ class m220916_205357_add_price_fields_to_game_table extends Migration
      */
     public function safeUp()
     {
-        $this->addColumn('{{%game}}', 'steam_price_initial', $this->smallInteger()->defaultValue(0)->after('steam_appid'));
-        $this->addColumn('{{%game}}', 'steam_price_final', $this->smallInteger()->defaultValue(0)->after('steam_price_initial'));
+        $this->addColumn('{{%game}}', 'steam_price_initial', $this->integer()->defaultValue(0)->after('steam_appid'));
+        $this->addColumn('{{%game}}', 'steam_price_final', $this->integer()->defaultValue(0)->after('steam_price_initial'));
 
         $this->createIndex('{{%game_steam_price_initial_index}}', '{{%game}}', 'steam_price_initial');
         $this->createIndex('{{%game_steam_price_final_index}}', '{{%game}}', 'steam_price_final');
